* fix some casts * introduce unary_op_type - cleaner than casting data voids to function ptrs
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@9282 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Zend/zend_execute_API.c
@@ -306,7 +306,7 @@ int call_user_function(HashTable *function_table, zval *object, zval *function_n
 		zend_ptr_stack_push(&EG(argument_stack), param);
 	}
 
-	zend_ptr_stack_push(&EG(argument_stack), (void *) param_count);
+	zend_ptr_stack_push(&EG(argument_stack), (void *) (long) param_count);
 
 	var_uninit(retval);
 	if (function_state.function->type == ZEND_USER_FUNCTION) {
